+++ R/add_phylopic.r
@@ -74,7 +74,7 @@
 #'   Note that png array objects can only be rotated by multiples of 90 degrees.
 #'   Also, outline colors do not currently work for png array objects.
 #' @importFrom ggplot2 annotate
-#' @importFrom lifecycle deprecated
+#' @importFrom lifecycle deprecated deprecate_warn is_present
 #' @export
 #' @examples \dontrun{
 #' # Put a silhouette behind a plot based on a taxonomic name